Update python.rst
I have read + tested the code and it works. For clarity I have added some extra line of code (e.g. how to test the function and how to test the class).

I have left some "???" for @rkdarst to fill in as I suppose he will be the one covering this chapter.